fix(ai): split diplomacy peace targets to satisfy file size lint (Refs #2509)
Extract stalled-expansion peace helpers into diplomacy_planner_peace_targets.dart
so diplomacy_planner.dart stays under the 1000 non-comment line repo gate.
